Make template generation less invasive
